+++ src/yaplox/stmt.py
@@ -32,6 +32,10 @@ class StmtVisitor(ABC):
         raise NotImplementedError
 
     @abstractmethod
+    def visit_return_stmt(self, stmt: "Return"):
+        raise NotImplementedError
+
+    @abstractmethod
     def visit_var_stmt(self, stmt: "Var"):
         raise NotImplementedError
 
@@ -93,6 +97,16 @@ class Print(Stmt):
     def accept(self, visitor: StmtVisitor):
         """ Create a accept method that calls the visitor. """
         return visitor.visit_print_stmt(self)
+
+
+class Return(Stmt):
+    def __init__(self, keyword: Token, value: Optional[Expr]):
+        self.keyword = keyword
+        self.value = value
+
+    def accept(self, visitor: StmtVisitor):
+        """ Create a accept method that calls the visitor. """
+        return visitor.visit_return_stmt(self)
 
 
 class Var(Stmt):