+++ tests/test_functions.py
@@ -183,3 +183,146 @@ class TestFunctions:
 
         captured = capsys.readouterr()
         assert captured.out == "1\n2\n3\n"
+
+    def test_return(self, capsys):
+        """
+        Test that a function can all itself.
+        This code example is also from the crafting interpreters book
+        """
+        statement = [
+            "fun add(a, b) {",
+            "    var result =  (a - 1) + (b - 1);",
+            "    return result + 2 ;",
+            "}",
+            "",
+            "print add(3, 5);",
+        ]
+        source = "\n".join(statement)
+        yaplox = Yaplox()
+        yaplox.run(source)
+
+        assert not yaplox.had_error
+        assert not yaplox.had_runtime_error
+
+        captured = capsys.readouterr()
+        assert captured.out == "8\n"
+
+    def test_recursive_fibonacci(self, capsys):
+        """
+        Example from crafting interpreters, a recursive Fibonacci function
+        """
+        statement = [
+            "fun fib(n) {",
+            "  if (n <= 1) return n;",
+            "  return fib(n - 2) + fib(n - 1);",
+            "}",
+            "",
+            "for (var i = 0; i < 20; i = i + 1) {",
+            "  print fib(i);",
+            "}",
+        ]
+        source = "\n".join(statement)
+        yaplox = Yaplox()
+        yaplox.run(source)
+
+        assert not yaplox.had_error
+        assert not yaplox.had_runtime_error
+
+        captured = capsys.readouterr()
+
+        # We expect the first 20 numbers:
+        # fmt: off
+        fibonacci = [
+            0, 1, 1, 2, 3,
+            5, 8, 13, 21, 34,
+            55, 89, 144, 233, 377,
+            610, 987, 1597, 2584, 4181,
+        ]
+        # fmt: on
+
+        expected_str_output = "\n".join([str(nr) for nr in fibonacci]) + "\n"
+        assert captured.out == expected_str_output
+
+    def test_return_nothing(self, capsys):
+        """
+        Test a function that returns nothing
+        """
+        statement = [
+            "fun return_nothing() { }",
+            "print return_nothing();",
+        ]
+        source = "\n".join(statement)
+        yaplox = Yaplox()
+        yaplox.run(source)
+
+        assert not yaplox.had_error
+        assert not yaplox.had_runtime_error
+
+        captured = capsys.readouterr()
+        assert captured.out == "nil\n"
+
+    def test_call_string(self, capsys):
+        """
+        Try to call a string. This must fail cleanly
+        """
+        source = '"ThisIsAstring"();'
+        yaplox = Yaplox()
+        yaplox.run(source)
+
+        assert not yaplox.had_error
+        assert yaplox.had_runtime_error
+
+        captured = capsys.readouterr()
+        assert captured.err == "Can only call functions and classes. in line [line1]\n"
+
+    def test_return_in_the_middle(self, capsys):
+        """
+        Test that we can return from a function in the middle of a loop.
+        It runs until 100, but will exit when n == 3, before printing n
+        """
+        statement = [
+            "fun count(n) {",
+            "  while (n < 100) {",
+            "    if (n == 3) return n; // <--",
+            "    print n;",
+            "    n = n + 1;",
+            "  }",
+            "}",
+            "",
+            "count(1);",
+        ]
+        source = "\n".join(statement)
+        yaplox = Yaplox()
+        yaplox.run(source)
+
+        assert not yaplox.had_error
+        assert not yaplox.had_runtime_error
+
+        captured = capsys.readouterr()
+        assert captured.out == "1\n2\n"
+
+    def test_empty_return(self, capsys):
+        """
+        This test looks like test_return_in_the_middle, but instead or returning n, it
+        will return nothing.
+        """
+        statement = [
+            "fun count(n) {",
+            "  while (n < 100) {",
+            "    if (n == 3) return; // <--",
+            "    print n;",
+            "    n = n + 1;",
+            "  }",
+            "}",
+            "",
+            "print count(1);",
+        ]
+        source = "\n".join(statement)
+        yaplox = Yaplox()
+        yaplox.run(source)
+
+        assert not yaplox.had_error
+        assert not yaplox.had_runtime_error
+
+        captured = capsys.readouterr()
+        assert captured.out == "1\n2\nnil\n"