+++ tools/generate_ast.py
@@ -42,6 +42,7 @@ class GenerateAst:
                 "Function   : Token name, List<Token> params, List<Stmt> body",
                 "If      : Expr condition, Stmt thenBranch, Optional[Stmt] elseBranch",
                 "Print      : Expr expression",
+                "Return     : Token keyword, Optional[Expr] value",
                 "Var        : Token name, Optional[Expr] initializer",
                 "While      : Expr condition, Stmt body",
             ],